Fixes bug when using time machine with monitor

+++ deemon/cmd/refresh.py
@@ -31,8 +31,9 @@ class Refresh:
         if self.time_machine:
             logger.info(f":: Time Machine active: {datetime.strftime(self.time_machine, '%b %d, %Y')}!")
             config.set('by_release_date', False)
-            self.db.remove_specific_releases({'tm_date': str(self.time_machine)})
-            self.db.commit()
+            if not self.waiting_for_refresh():
+                self.db.remove_specific_releases({'tm_date': str(self.time_machine)})
+                self.db.commit()
 
     @staticmethod
     def debugger(message: str, payload = None):
